* List_GEN.m ([TARGET -describe:]): Cast getField return to int.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@4537 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/collections/List_GEN.m
@@ -225,7 +225,7 @@ PHASE(UsingOnly)
   [super describe: outputCharStream];
 #if MLINKS
   sprintf (buffer, "> internal links at offset: %d\n",
-           getField (bits, IndexFromMemberLoc_Shift, IndexFromMemberLoc_Mask) +
+           (int) getField (bits, IndexFromMemberLoc_Shift, IndexFromMemberLoc_Mask) +
            IndexFromMemberLoc_Min);
   [outputCharStream catC: buffer];
 #endif
